calibration procedure has higher accuracy now

+++ src/hardware/calibration.cpp
@@ -133,7 +133,7 @@ uint16_t _calibration_do_scale_calibration(Waveformer& wf) {
 }
 
 uint16_t _calibration_calc_vo_scale(uint16_t one_volt, uint16_t five_volts) {
-    return static_cast<uint16_t>((1 << 19) / (one_volt - five_volts));
+    return static_cast<uint16_t>((1 << 19) / static_cast<uint32_t>(one_volt - five_volts));
 }
 
 uint16_t _calibration_calc_vo_offset(uint16_t vo_scale, uint16_t zero_volts) {
